Added worker threads computing blur.
Blur is currently not working - the outputted image file is a 16x16 black tile. However, the workers are 'working' and the skeleton seems okay.

The program works by reading 3 lines of the file at a time and distributing these to the workers, that each compute a single pixel, which is returned. The program cannot have deadlock as each worker is sent data in sequence, then each response is read in the same sequence.