Additional shutdown procedures.
The button LEDs are now turned off appropriately when shutdown is pressed as the first button. This is done by sending the shutdown message from the buttonlistener to the visualiser immediately.

The DataInStream thread will now terminate gracefully when shutdown is pressed as the first button (i.e. before processing in the distributor begins).